requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic==2.10.0
pydantic-settings==2.7.1
passlib[bcrypt]
python-jose[cryptography]
email-validator
cloudinary
aiofiles
fastapi-limiter
redis
psycopg2-binary
alembic
python-multipart
sphinx
sphinx-autodoc-typehints
sphinx-rtd-theme
